[FLINK-12755][hive] move tests for data type mappings between Flink and Hive into its own test class
