+++ flink-connectors/flink-connector-hive/src/test/java/org/apache/flink/table/catalog/hive/HiveCatalogGenericMetadataTest.java
@@ -18,25 +18,11 @@
 
 package org.apache.flink.table.catalog.hive;
 
-import org.apache.flink.table.api.DataTypes;
-import org.apache.flink.table.api.TableSchema;
 import org.apache.flink.table.catalog.CatalogPartition;
-import org.apache.flink.table.catalog.CatalogTable;
-import org.apache.flink.table.catalog.CatalogTableImpl;
 import org.apache.flink.table.catalog.CatalogTestBase;
-import org.apache.flink.table.catalog.exceptions.CatalogException;
-import org.apache.flink.table.types.DataType;
-import org.apache.flink.table.types.logical.BinaryType;
-import org.apache.flink.table.types.logical.VarBinaryType;
 
-import org.apache.hadoop.hive.common.type.HiveChar;
-import org.apache.hadoop.hive.common.type.HiveVarchar;
 import org.junit.BeforeClass;
 import org.junit.Test;
-
-import java.util.Arrays;
-
-import static org.junit.Assert.assertEquals;
 
 /**
  * Test for HiveCatalog on generic metadata.
@@ -47,121 +33,6 @@ public class HiveCatalogGenericMetadataTest extends CatalogTestBase {
 	public static void init() {
 		catalog = HiveTestUtils.createHiveCatalog();
 		catalog.open();
-	}
-
-	// ------ TODO: Move data types tests to its own test class as it's shared between generic metadata and hive metadata
-	// ------ data types ------
-
-	@Test
-	public void testDataTypes() throws Exception {
-		// TODO: the following Hive types are not supported in Flink yet, including MAP, STRUCT
-		DataType[] types = new DataType[] {
-			DataTypes.TINYINT(),
-			DataTypes.SMALLINT(),
-			DataTypes.INT(),
-			DataTypes.BIGINT(),
-			DataTypes.FLOAT(),
-			DataTypes.DOUBLE(),
-			DataTypes.BOOLEAN(),
-			DataTypes.STRING(),
-			DataTypes.BYTES(),
-			DataTypes.DATE(),
-			DataTypes.TIMESTAMP(),
-			DataTypes.CHAR(HiveChar.MAX_CHAR_LENGTH),
-			DataTypes.VARCHAR(HiveVarchar.MAX_VARCHAR_LENGTH),
-			DataTypes.DECIMAL(5, 3)
-		};
-
-		verifyDataTypes(types);
-	}
-
-	@Test
-	public void testNonExactlyMatchedDataTypes() throws Exception {
-		DataType[] types = new DataType[] {
-			DataTypes.BINARY(BinaryType.MAX_LENGTH),
-			DataTypes.VARBINARY(VarBinaryType.MAX_LENGTH)
-		};
-
-		CatalogTable table = createCatalogTable(types);
-
-		catalog.createDatabase(db1, createDb(), false);
-		catalog.createTable(path1, table, false);
-
-		Arrays.equals(
-			new DataType[] {DataTypes.BYTES(), DataTypes.BYTES()},
-			catalog.getTable(path1).getSchema().getFieldDataTypes());
-	}
-
-	@Test
-	public void testCharTypeLength() throws Exception {
-		DataType[] types = new DataType[] {
-			DataTypes.CHAR(HiveChar.MAX_CHAR_LENGTH + 1)
-		};
-
-		exception.expect(CatalogException.class);
-		exception.expectMessage("HiveCatalog doesn't support char type with length of '256'. The maximum length is 255");
-		verifyDataTypes(types);
-	}
-
-	@Test
-	public void testVarCharTypeLength() throws Exception {
-		DataType[] types = new DataType[] {
-			DataTypes.VARCHAR(HiveVarchar.MAX_VARCHAR_LENGTH + 1)
-		};
-
-		exception.expect(CatalogException.class);
-		exception.expectMessage("HiveCatalog doesn't support varchar type with length of '65536'. The maximum length is 65535");
-		verifyDataTypes(types);
-	}
-
-	@Test
-	public void testComplexDataTypes() throws Exception {
-		DataType[] types = new DataType[]{
-			DataTypes.ARRAY(DataTypes.DOUBLE()),
-			DataTypes.MAP(DataTypes.FLOAT(), DataTypes.BIGINT()),
-			DataTypes.ROW(
-				DataTypes.FIELD("0", DataTypes.BOOLEAN()),
-				DataTypes.FIELD("1", DataTypes.BOOLEAN()),
-				DataTypes.FIELD("2", DataTypes.DATE())),
-
-			// nested complex types
-			DataTypes.ARRAY(DataTypes.ARRAY(DataTypes.INT())),
-			DataTypes.MAP(DataTypes.STRING(), DataTypes.MAP(DataTypes.STRING(), DataTypes.BIGINT())),
-			DataTypes.ROW(
-				DataTypes.FIELD("3", DataTypes.ARRAY(DataTypes.DECIMAL(5, 3))),
-				DataTypes.FIELD("4", DataTypes.MAP(DataTypes.TINYINT(), DataTypes.SMALLINT())),
-				DataTypes.FIELD("5", DataTypes.ROW(DataTypes.FIELD("3", DataTypes.TIMESTAMP())))
-			)
-		};
-
-		verifyDataTypes(types);
-	}
-
-	private CatalogTable createCatalogTable(DataType[] types) {
-		String[] colNames = new String[types.length];
-
-		for (int i = 0; i < types.length; i++) {
-			colNames[i] = String.format("%s_%d", types[i].toString().toLowerCase(), i);
-		}
-
-		TableSchema schema = TableSchema.builder()
-			.fields(colNames, types)
-			.build();
-
-		return new CatalogTableImpl(
-			schema,
-			getBatchTableProperties(),
-			TEST_COMMENT
-		);
-	}
-
-	private void verifyDataTypes(DataType[] types) throws Exception {
-		CatalogTable table = createCatalogTable(types);
-
-		catalog.createDatabase(db1, createDb(), false);
-		catalog.createTable(path1, table, false);
-
-		assertEquals(table.getSchema(), catalog.getTable(path1).getSchema());
 	}
 
 	// ------ partitions ------
